CH1 frequency changes
When the CPU writes to the frequency it updates the lower or upper bits
of freq_divider. I added the byte enable signals to eProcReg_gba for
that.

I changed the frequency counter a bit. Instead of counting from 0 to
(2048 - freq_divider). It just counts to 2047 starting from the value in
freq_divider. When the Start bit is set or when freq_cnt is 2047 then it
reloads freq_divider into freq_cnt.