Copy BGX/Y at the start of the line if written
If BGX/Y is written in the middle of the line then the next line should
use the new value.

Actually the value is already latched at the start of drawing in
gba_drawer_mode* but the ref2_x/y values are incremented at the start of
HBlank and that would overwrite the newly written value.